chip8raw: first semi-usable version
This is the first semi-usable version of the c8r assembler mode. Dynamic
symbol resolution is not yet implemented, but is hopefully coming soon.
Right now any attempts to invoke a symbol panics. However, you could
technically use this (with something like to-be-working xas-asm) to
assemble actual c8r.

Cleanup is WIP. First hopeful objective is to clear out the mess of gim!
calls in get_instruction(). Cleaning out some of the obtain functions
and functionizing the byte formatting instructions is to-be-considered.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>