chip8raw: phase 1 of symbols complete
All c8r instructions now have fully working symbols! This is honestly
great. This is, honestly, incredible progress. However, something has
come up...

Phase 2 and 3 will likely need to be intermingled. Going over the rawbin
code - as well as Platform and BBU::outs - shows that this can't really
be worked with, at least not well, without making the necessary changes
to general architecture. This means entirely scrapping the current outs
system, scrapping *almost all* of Platform, and significantly
overhauling the Lexer. This all but guarantees significant changes to
EAF as well.

I don't think, at this point, that a commit in any form is possible - at
least with keeping the status of "it needs to compile to be a valid
commit", especially considering future bisects - until both phases 2 and
3 are complete.

I didn't originally think that v0.1 would be publishable. I still am not
sure. But with the amount of overhaul that's being put in here... Maybe
cleanup doesn't happen in 0.1 per se, maybe it comes in 0.2 or 0.3, but
a pre-1.0 publication is almost certainly possible.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>