bbu: symbols phase 2 mostly done
Symbols Phase 2 is almost completely done. The general output interface
and EAF still need to be reworked; they are currently unavailable, so
upgrading to this commit will not work.

From now on, all commits will have a changelog. Here's the first!
- bbu: fulfill_symbol has had parameter `s`'s type changed from `T` to
`&T`. This is to prevent unnecessary code and conversions.
- ArchInstruction elements in C8r now derive Clone. This is NOT
NECESSARY as ArchInstruction itself can naturally not hold Clone as it
is not inherently sized. See #4.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>