cleanup: kill all dead code
This patch removes, as part of the pre-release cleanup process, all dead
code segments lying around. It also updates some TODOs for relevance;
this is not guaranteed to be complete, however. Per the commit, it is
estimated that this patch saves 131 LOC from counts, reducing code size
a decent amount. Some old/irrelevant comments were removed. Commenting
practices definitely need to be improved, and documentation/standards
need to be formalized, but that's hopefully for 0.2/0.3 - this is just
basic cleaning.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>