lexer: remove LexOp::Macro
This patch finally gets rid of the Macro enum-part in LexOperation.
There's still other associated cleanup that needs to be done - and we
really need to consider getting rid of LexOp altogether - but this is a
start which should shave some bloat off. The code has just been
commented out for now in case it needs to be reenabled in the future.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>